Fixes issue where span ID is bigger than long in GAE
App Engine generates span IDs that are longer than a long.
Given that Stackdriver Trace doesn't support trace joins, the incoming
span ID is useless anyway.
So, to save the time of parsing and creating a new long object, we don't
do that and, instead, set 1L as the span ID value, which Brave will
always regenerate anyways.