Refactors StackdriverTracePropagation so that B3 can be reconfigured (#174)
For a few releases, Brave has been able to configure `B3` differently
per request type. For example, messaging configured with single headers.

This makes it possible for `StackdriverTracePropagation` to honor that,
notably by fixing an incorrect package name! For example, depending on
the version of sleuth in use, this can now change the policy for which
headers are used for local propagation.

See https://github.com/spring-cloud/spring-cloud-sleuth/issues/1643